Bug 1353650 - Part 4: Update to ICU 59. rs=Waldo
UltraBlame original commit: 3b59d9b42f78091509507be1eeb254012d879a61

+++ intl/icu/source/i18n/nfrs.h
@@ -88,7 +88,7 @@ private:
 int64_t util64_fromDouble(double d);
 
 
-int64_t util64_pow(int32_t radix, uint32_t exponent);
+int64_t util64_pow(int32_t radix, uint16_t exponent);
 
 
 uint32_t util64_tou(int64_t n, UChar* buffer, uint32_t buflen, uint32_t radix = 10, UBool raw = FALSE);
