+++ intl/icu/source/i18n/nfsubs.h
@@ -91,7 +91,7 @@ public:
 
 
 
-    virtual void setDivisor(int32_t radix, int32_t exponent, UErrorCode& status);
+    virtual void setDivisor(int32_t radix, int16_t exponent, UErrorCode& status);
